+++ pos/src/main/java/com/fullstack/pos/dto/ItemTransaccionDTO.java
@@ -1,6 +1,7 @@
 package com.fullstack.pos.dto;
 
 import lombok.*;
+import java.math.BigDecimal;
 
 @Data
 @AllArgsConstructor
@@ -8,10 +9,10 @@ import lombok.*;
 @Builder
 public class ItemTransaccionDTO {
     private Long id;
-    private Long transaccionId; // Solo el ID, no el objeto completo
+    private Long transaccionId;
     private Long productoId;
     private String descripcion;
     private Integer cantidad;
-    private Double precioUnitario;
-    private Double subTotal;
+    private BigDecimal precioUnitario;
+    private BigDecimal subTotal;
 }
